gemini-vision-tutor/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-generativeai>=0.8.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
pydantic==2.7.1
Pillow==10.3.0
